♻️(backend) split BBB classroom API tests
Our test file became too long.
We deceded to split them to dedicated files, reflecting our API
endpoints.
