+++ src/backend/marsha/bbb/tests/api/classroom/test_retrieve.py
@@ -1,0 +1,160 @@
+"""Tests for the classroom retrieve API."""
+from datetime import datetime, timedelta
+import json
+from unittest import mock
+import zoneinfo
+
+from django.test import TestCase, override_settings
+
+from marsha.bbb import serializers
+from marsha.bbb.factories import ClassroomFactory
+from marsha.core.simple_jwt.factories import (
+    InstructorOrAdminLtiTokenFactory,
+    StudentLtiTokenFactory,
+)
+from marsha.core.tests.utils import reload_urlconf
+
+
+# We don't enforce arguments documentation in tests
+# pylint: disable=unused-argument
+
+
+@override_settings(BBB_API_ENDPOINT="https://10.7.7.1/bigbluebutton/api")
+@override_settings(BBB_API_SECRET="SuperSecret")
+@override_settings(BBB_ENABLED=True)
+class ClassroomRetrieveAPITest(TestCase):
+    """Test for the Classroom retrieve API."""
+
+    maxDiff = None
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+
+        # Force URLs reload to use BBB_ENABLED
+        reload_urlconf()
+
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_classroom_fetch_student(self, mock_get_meeting_infos):
+        """A student should be allowed to fetch a classroom."""
+        classroom = ClassroomFactory()
+        mock_get_meeting_infos.return_value = {
+            "returncode": "SUCCESS",
+            "running": "true",
+        }
+
+        jwt_token = StudentLtiTokenFactory(resource=classroom)
+
+        response = self.client.get(
+            f"/api/classrooms/{classroom.id!s}/",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+        self.assertDictEqual(
+            {
+                "id": str(classroom.id),
+                "infos": {"returncode": "SUCCESS", "running": "true"},
+                "lti_id": str(classroom.lti_id),
+                "title": classroom.title,
+                "description": classroom.description,
+                "started": False,
+                "ended": False,
+                "meeting_id": str(classroom.meeting_id),
+                "welcome_text": classroom.welcome_text,
+                "playlist": {
+                    "id": str(classroom.playlist.id),
+                    "title": classroom.playlist.title,
+                    "lti_id": classroom.playlist.lti_id,
+                },
+                "starting_at": None,
+                "estimated_duration": None,
+            },
+            content,
+        )
+
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_classroom_fetch_student_scheduled(self, mock_get_meeting_infos):
+        """A student should be allowed to fetch a scheduled classroom."""
+        now = datetime(2018, 8, 8, tzinfo=zoneinfo.ZoneInfo("Europe/Paris"))
+
+        # set microseconds to 0 to compare date surely as serializer truncate them
+        starting_at = (now + timedelta(hours=3)).replace(microsecond=0)
+        estimated_duration = timedelta(seconds=60)
+
+        classroom = ClassroomFactory(
+            starting_at=starting_at, estimated_duration=estimated_duration
+        )
+        mock_get_meeting_infos.return_value = {
+            "returncode": "SUCCESS",
+            "running": "true",
+        }
+
+        jwt_token = StudentLtiTokenFactory(resource=classroom)
+
+        response = self.client.get(
+            f"/api/classrooms/{classroom.id!s}/",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+        self.assertDictEqual(
+            {
+                "id": str(classroom.id),
+                "infos": {"returncode": "SUCCESS", "running": "true"},
+                "lti_id": str(classroom.lti_id),
+                "title": classroom.title,
+                "description": classroom.description,
+                "started": False,
+                "ended": False,
+                "meeting_id": str(classroom.meeting_id),
+                "welcome_text": classroom.welcome_text,
+                "playlist": {
+                    "id": str(classroom.playlist.id),
+                    "title": classroom.playlist.title,
+                    "lti_id": classroom.playlist.lti_id,
+                },
+                "starting_at": "2018-08-08T01:00:00Z",
+                "estimated_duration": "00:01:00",
+            },
+            content,
+        )
+
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_classroom_fetch_instructor(self, mock_get_meeting_infos):
+        """An instructor should be able to fetch a classroom."""
+        classroom = ClassroomFactory()
+        mock_get_meeting_infos.return_value = {
+            "returncode": "SUCCESS",
+            "running": "true",
+        }
+
+        jwt_token = InstructorOrAdminLtiTokenFactory(resource=classroom)
+
+        response = self.client.get(
+            f"/api/classrooms/{classroom.id!s}/",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+        self.assertDictEqual(
+            {
+                "id": str(classroom.id),
+                "infos": {"returncode": "SUCCESS", "running": "true"},
+                "lti_id": str(classroom.lti_id),
+                "title": classroom.title,
+                "description": classroom.description,
+                "started": False,
+                "ended": False,
+                "meeting_id": str(classroom.meeting_id),
+                "welcome_text": classroom.welcome_text,
+                "playlist": {
+                    "id": str(classroom.playlist.id),
+                    "title": classroom.playlist.title,
+                    "lti_id": classroom.playlist.lti_id,
+                },
+                "starting_at": None,
+                "estimated_duration": None,
+            },
+            content,
+        )
