+++ src/backend/marsha/bbb/tests/api/classroom/test_service_join.py
@@ -1,0 +1,144 @@
+"""Tests for the classroom API."""
+import json
+from unittest import mock
+from urllib.parse import quote_plus
+
+from django.test import TestCase, override_settings
+
+from marsha.bbb import api, serializers
+from marsha.bbb.factories import ClassroomFactory
+from marsha.core import factories as core_factories
+from marsha.core.simple_jwt.factories import (
+    InstructorOrAdminLtiTokenFactory,
+    StudentLtiTokenFactory,
+)
+from marsha.core.tests.utils import reload_urlconf
+
+
+# We don't enforce arguments documentation in tests
+# pylint: disable=unused-argument
+
+
+@override_settings(BBB_API_ENDPOINT="https://10.7.7.1/bigbluebutton/api")
+@override_settings(BBB_API_SECRET="SuperSecret")
+@override_settings(BBB_ENABLED=True)
+class ClassroomServiceJoinAPITest(TestCase):
+    """Test for the Classroom API."""
+
+    maxDiff = None
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+
+        # Force URLs reload to use BBB_ENABLED
+        reload_urlconf()
+
+    @mock.patch.object(api, "join")
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_bbb_join_classroom_anonymous(
+        self, mock_get_meeting_infos, mock_join_request
+    ):
+        """An anonymous should not be able to join a classroom."""
+        classroom = ClassroomFactory()
+
+        response = self.client.patch(
+            f"/api/classrooms/{classroom.id}/join/",
+        )
+        self.assertEqual(response.status_code, 401)
+        mock_get_meeting_infos.assert_not_called()
+        mock_join_request.assert_not_called()
+
+    @mock.patch.object(api, "join")
+    def test_api_bbb_join_classroom_user_logged_in(self, mock_join_request):
+        """A logged in user should not be able to join a classroom."""
+        user = core_factories.UserFactory(
+            first_name="Jane", last_name="Doe", email="jane.doe@example.com"
+        )
+        classroom = ClassroomFactory()
+        self.client.force_login(user)
+
+        response = self.client.patch(f"/api/classrooms/{classroom.id}/join/")
+        self.assertEqual(response.status_code, 401)
+        mock_join_request.assert_not_called()
+
+    def test_api_bbb_join_student(self):
+        """Joining a classroom as student should return an attendee classroom url."""
+        classroom = ClassroomFactory(
+            meeting_id="21e6634f-ab6f-4c77-a665-4229c61b479a",
+            title="Classroom 1",
+            attendee_password="9#R1kuUl3R",
+            moderator_password="0$C7Aaz0o",
+        )
+
+        jwt_token = StudentLtiTokenFactory(
+            resource=classroom,
+            consumer_site="consumer_site",
+            user__id="user_id",
+        )
+
+        response = self.client.patch(
+            f"/api/classrooms/{classroom.id}/join/",
+            data=json.dumps({"fullname": "John Doe"}),
+            content_type="application/json",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertIn(
+            "https://10.7.7.1/bigbluebutton/api/join?"
+            f"fullName=John+Doe&meetingID={classroom.meeting_id}&"
+            f"password={quote_plus(classroom.attendee_password)}&"
+            "userID=consumer_site_user_id&redirect=true",
+            response.data.get("url"),
+        )
+
+    def test_api_bbb_join_instructor(self):
+        """Joining a classroom as instructor should return a moderator classroom url."""
+        classroom = ClassroomFactory(
+            meeting_id="21e6634f-ab6f-4c77-a665-4229c61b479a",
+            title="Classroom 1",
+            attendee_password="9#R1kuUl3R",
+            moderator_password="0$C7Aaz0o",
+        )
+
+        jwt_token = InstructorOrAdminLtiTokenFactory(
+            resource=classroom,
+            consumer_site="consumer_site",
+            user__id="user_id",
+        )
+
+        response = self.client.patch(
+            f"/api/classrooms/{classroom.id}/join/",
+            data=json.dumps({"fullname": "John Doe"}),
+            content_type="application/json",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertIn(
+            "https://10.7.7.1/bigbluebutton/api/join?"
+            f"fullName=John+Doe&meetingID={classroom.meeting_id}&"
+            f"password={quote_plus(classroom.moderator_password)}&"
+            "userID=consumer_site_user_id&redirect=true",
+            response.data.get("url"),
+        )
+
+    def test_api_bbb_join_instructor_no_fullname(self):
+        """Joining a classroom without fullname parameter should return a 422."""
+        classroom = ClassroomFactory(
+            meeting_id="21e6634f-ab6f-4c77-a665-4229c61b479a",
+            title="Classroom 1",
+            attendee_password="9#R1kuUl3R",
+            moderator_password="0$C7Aaz0o",
+        )
+
+        jwt_token = InstructorOrAdminLtiTokenFactory(resource=classroom)
+
+        response = self.client.patch(
+            f"/api/classrooms/{classroom.id}/join/",
+            HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
+        )
+        self.assertEqual(response.status_code, 400)
+        self.assertDictEqual(
+            {"message": "missing fullname parameter"},
+            response.data,
+        )
